# openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Excel GPT Middleware API",
    "version": "1.1.0",
    "description": "Express API middleware for Microsoft Graph Excel with name-based resolution, recursive search, Excel engine formatting, intelligent find/replace, and rename operations.\nNotes:\n- All Excel and rename endpoints are mounted under /api/excel. Health endpoints are under /health.\n- Name-vs-ID: All endpoints accept human-friendly names (driveName, folderName, fileName, sheetName) with automatic name-to-ID resolution and backward compatibility for IDs.\n- Path disambiguation: When duplicate file names exist, you can pass itemPath (e.g., '/Folder/Subfolder/My.xlsx').\n- Error envelope is consistent: { status: \"success\" | \"error\" | \"partial_success\" | \"preview\" | \"multiple_matches\" | \"no_matches\", ... }. Success responses include status: \"success\" and may include data."
  },
  "servers": [
    {
      "url": "https://excel-middleware.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/excel/validate-formula": {
      "post": {
        "summary": "Validate Excel formula syntax",
        "operationId": "validateFormula",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateFormulaRequest" } } } },
        "responses": {
          "200": { "description": "Validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateFormulaSuccess" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
    "/api/excel/worksheet-info": {
      "get": {
        "summary": "Worksheet structure and metadata",
        "operationId": "worksheetInfo",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site ID override" },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site URL override" },
          { "in": "query", "name": "sharepointHostname", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint hostname override (tenant.sharepoint.com)" },
          { "in": "query", "name": "sharepointSiteName", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint site name override (e.g., Finance)" },
          { "in": "query", "name": "driveName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemPath", "schema": { "type": "string" } },
          { "in": "query", "name": "sheetName", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Worksheet info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "EnvelopeSuccessWithData": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success"], "example": "success" },
          "data": {}
        },
        "required": ["status"]
      },
      "EnvelopeError": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["error"], "example": "error" },
          "error": { "oneOf": [ { "type": "string" }, { "type": "object", "additionalProperties": true } ] }
        },
        "required": ["status", "error"]
      },
...
      "ValidateFormulaRequest": {
        "type": "object",
        "properties": {
          "driveName": { "type": "string" },
          "itemName": { "type": "string" },
          "itemPath": { "type": "string" },
          "sheetName": { "type": "string" },
          "formula": {
            "type": "object",
            "properties": {
              "expression": { "type": "string" },
              "targetCell": { "type": "string" },
              "overwrite": { "type": "boolean" }
            },
            "required": ["expression"]
          }
        },
        "required": ["itemName", "formula"]
      },
      "ValidateFormulaSuccess": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success"], "example": "success" },
          "valid": { "type": "boolean" },
          "message": { "type": "string" }
        },
        "required": ["status", "valid"]
      },
...
    }
  }
}